refactor: Initialize.tsx 구 상태관리 모두 제거
