--- src/components/Initialize/index.ts
@@ -1,3 +1,0 @@
-import Initialize from './Initialize';
-
-export { Initialize };
