+++ src/components/index.ts
@@ -2,7 +2,7 @@ import ButtonOfPages from './ButtonOfPage';
 import Datepicker from './Datepicker';
 import CheckableButton from './CheckableButton';
 import CheckAdmin from './CheckAdmin';
-import { Initialize } from './Initialize';
+import Initialize from './Initialize';
 import ListCell from './ListCell';
 import MenuButton from './MenuButton';
 import Selectbox from './Selectbox';
